requirements.txt
pyhive[hive]==0.7.0
thrift==0.20.0
thrift-sasl==0.4.3
PyYAML==6.0.2

